Lambda.Parser

Description

Parses lambda code.

Documentation

term :: Parser (Expr Integer)

Parses a lambda term, which refers to the term instantiated by the lambda at a certain depth (0-indexed).

application :: Parser (Expr Integer)

Parses the application of two expressions, which must be wrapped in brackets.

lambda :: Parser (Expr Integer)

Parses a lambda, which consists of a dot and an expression.

expression :: Parser (Expr Integer)

Parses an expression, which is either a(n): term; lambda; application.

input :: ExceptT String IO (Expr Integer)

See inputGenerator.