Parsers

Description

Contains helpers for parsers and for running parsers.

Documentation

type Parser = Parsec Void String

The default type used for a parser.

symbol :: Parser a -> Parser a

Consumes all spaces around a parser.

sSymbol :: String -> Parser String

Consumes a string, whilst remaining ignorant to spaces around it.

brackets :: Parser a -> Parser a

Wraps a parser in brackets that ignore whitespace.

inputGenerator :: String -> Parser a -> ExceptT String IO a

Creates an 'ExceptT String IO a', that reads and parses a file, whilst displaying errors accordingly. This does not handle IO errors.