Transpiler.Parser

Description

Parses Iocularia code.

Documentation

term :: Parser String

Parses a term, which must consist of at least one letterChar. This parser is able to backtrack.

lhs :: Parser LHS

Parses the left hand side of an equation, consisting more than 1 space separated terms, the first of which being the name of the variable.

rhs :: Parser RHS

Parses the right hand side of an equation, which can either be a term, or an application of two '@rhs'. Application is left associative.

equation :: Parser Equation

Parses an Equation.

program :: Parser [Equation]

Parses a program: many Equations.

input :: ExceptT String IO [Equation]

See inputGenerator.